Дашборд v4: SPA на Vite + React + Tailwind v4 + Recharts
frontend/ — логин по JWT, выбор устройства и три блока поверх API v3:

- KPI-карты: Rth (последняя честная точка + 7-дневный уровень), T_ambient
  с бейджем уверенности калибровки (подпись честная: «опорный idle-режим
  устройства», не комната), Health Score с цветовыми зонами §5.5 и
  стейтом «Идёт калибровка данных…» пока эпоха sparse
- скаттер истории Rth: точки ≥ per-device гейта качества (гейт показан
  в подписи), пунктирная линия базлайна эпохи, тултип с P/стратой/
  качеством, переключатели cpu/gpu и 7/30/90 дней
- карточка CUSUM-предложения с выбором вида работ и «Подтвердить» =
  POST /maintenance c performed_at дня ступеньки — попадает в ±3-дневное
  окно закрытия предложения

Типы src/types.ts зеркалят schemas/devices.py — формы из хендоффа были
искажены (domain_name, details, quality_gate-объект), истина — схемы.
CORS бэкенду не добавлен: dev-прокси Vite (THERMAL_API_TARGET для запуска
в докере на сети infra_default; node_modules — в named volume, не в
OneDrive). UI-примитивы в стилистике shadcn без CLI-зависимости.

Проверено e2e в браузере на живом стеке с данными, прогнанными через
реальный пайплайн (update_trends сам дал базлайн, скор 60 и CUSUM-
предложение): подтверждение из UI закрыло предложение, перезапустило
эпоху — базлайн на графике переехал 1.0 → 1.15, скор 60 → 90. Сборка
tsc+vite зелёная; попутно — float() на step_pct (np.float64 протекал
в structlog и note предложения).

+++ backend/app/analytics/jobs.py
@@ -323,7 +323,8 @@ async def update_trends_for_device(
             # §5.5: чейнджпойнт ⇒ предложение подтвердить событие обслуживания.
             # Эпоху НЕ режет (kind не в EPOCH_RESET_KINDS) — псевдо-граница
             # тренда уже обеспечена сегментом seg; базлайн ждёт подтверждения.
-            step_pct = 100.0 * (medians[seg] - medians[seg - 1]) / medians[seg - 1]
+            # float(): иначе np.float64 протекает в structlog-вывод и note
+            step_pct = float(100.0 * (medians[seg] - medians[seg - 1]) / medians[seg - 1])
             await _suggest_regime_change(
                 conn, device_id, domain, stratum,
                 daily[seg]["day"], step_pct, _tz(tz_name),
